+++ src/main/kotlin/com/zepben/evolve/cim/iec61968/common/TownDetail.kt
@@ -13,5 +13,17 @@ package com.zepben.evolve.cim.iec61968.common
  * @property name Town name.
  * @property stateOrProvince Name of the state or province.
  */
-data class TownDetail(val name: String, val stateOrProvince: String)
+data class TownDetail(
+    val name: String? = null,
+    val stateOrProvince: String? = null
+) {
+
+    /**
+     * Check to see if all fields of this [TownDetail] are null or empty
+     *
+     * @return true if all fields are null or empty, otherwise false
+     */
+    fun allFieldsNullOrEmpty(): Boolean = name.isNullOrEmpty() && stateOrProvince.isNullOrEmpty()
+
+}