+++ src/test/kotlin/com/zepben/evolve/cim/iec61968/common/TownDetailTest.kt
@@ -22,7 +22,7 @@ internal class TownDetailTest {
 
     @Test
     internal fun constructorCoverage() {
-        assertThat(TownDetail("", ""), notNullValue())
+        assertThat(TownDetail(), notNullValue())
     }
 
     @Test
@@ -32,4 +32,14 @@ internal class TownDetailTest {
         assertThat(townDetail.name, equalTo("name"))
         assertThat(townDetail.stateOrProvince, equalTo("stateOrProvince"))
     }
+
+    @Test
+    internal fun testAllFieldsNullOrEmpty() {
+        assertThat(TownDetail().allFieldsNullOrEmpty(), equalTo(true))
+        assertThat(TownDetail("", "").allFieldsNullOrEmpty(), equalTo(true))
+
+        assertThat(TownDetail(name = "value").allFieldsNullOrEmpty(), equalTo(false))
+        assertThat(TownDetail(stateOrProvince = "value").allFieldsNullOrEmpty(), equalTo(false))
+    }
+
 }